fix: KEEP-354 isolate SSRF guard so drizzle.config.ts stays import-clean
The KEEP-354 SSRF commit added `import { isBlockedIp } from
"@/lib/safe-fetch"` to lib/db/connection-utils.ts. safe-fetch.ts
imports `server-only`, so connection-utils now transitively pulls
that in. drizzle.config.ts loads connection-utils (for getDatabaseUrl)
when drizzle-kit runs from the CLI, which is not a Server Component
context, so module evaluation throws:

  Error: This module cannot be imported from a Client Component
  module. It should only be used from a Server Component.

That blocks the e2e-vitest-ephemeral CI job at the Setup database
step, which in turn skips the deploy job. Three staging-branch
merges in a row failed to deploy because of this:
- PR #999 (the migration that introduced the bug)
- PR #998 (docs follow-up, hit the same path)
- PR #1004 (unrelated llms.txt feature)

Fix: extract the SSRF guard into its own file
(lib/db/connection-host-guard.ts) which keeps the `server-only`
import. connection-utils.ts is restored to its original (CLI-safe)
shape. test-connection.ts and the unit test update their imports to
the new module. No behaviour change; same guard fires on the same
inputs.

Sanity-checked by running `import("./drizzle.config.ts")` directly
under tsx -- previously threw the server-only error, now resolves
to the config object.
