+++ lib/db/connection-utils.ts
@@ -1,13 +1,10 @@
 /**
  * Database connection utilities
  *
- * Handles URL encoding of database credentials with special characters
+ * Handles URL encoding of database credentials with special characters.
+ * Imports must stay free of `server-only` (transitively) because
+ * `drizzle.config.ts` loads this file at CLI time.
  */
-
-import type { LookupAddress } from "node:dns";
-import { promises as dns } from "node:dns";
-import { isIP } from "node:net";
-import { isBlockedIp } from "@/lib/safe-fetch";
 
 /**
  * Protocol prefix for PostgreSQL connection strings
@@ -158,65 +155,6 @@ export type DatabaseConnectionConfig = {
   password?: string;
   database?: string;
 };
-
-/**
- * Reject connection-test attempts against private / loopback / link-local
- * destinations. Used by the integration-test endpoint, which dials a
- * user-supplied host over raw TCP via the `postgres` package -- a path that
- * does not go through `safe-fetch.ts` (HTTP-only). Always-on (no env flag):
- * there is no legitimate reason for an `/api/integrations/test` request to
- * resolve to a non-public address.
- *
- * Caveat: there is a small TOCTOU window between this DNS check and the
- * subsequent TCP connect during which an attacker-controlled domain could
- * rebind. Closing that window requires resolving once and dialling the
- * resolved IP directly, which is out of scope for this commit (would mean
- * rewriting how the postgres client is invoked). The TOCTOU surface for
- * postgres TCP is much narrower than HTTP DNS rebinding (no scriptable
- * client behind the connection), so we accept it here.
- */
-export async function assertHostIsPublic(host: string): Promise<void> {
-  const trimmed = host.trim();
-  if (trimmed === "") {
-    throw new Error("Host is required");
-  }
-
-  if (isIP(trimmed) !== 0) {
-    const verdict = isBlockedIp(trimmed);
-    if (verdict.blocked) {
-      throw new Error("Host is not allowed: must resolve to a public address");
-    }
-    return;
-  }
-
-  let addresses: LookupAddress[];
-  try {
-    addresses = await dns.lookup(trimmed, { all: true });
-  } catch {
-    throw new Error("Host could not be resolved");
-  }
-
-  for (const addr of addresses) {
-    const verdict = isBlockedIp(addr.address);
-    if (verdict.blocked) {
-      throw new Error("Host is not allowed: must resolve to a public address");
-    }
-  }
-}
-
-/**
- * Extract the hostname from a postgres connection URL.
- *
- * Returns null if the URL is malformed; callers should treat that as a
- * validation failure.
- */
-export function extractHostFromConnectionString(url: string): string | null {
-  try {
-    return new URL(url).hostname || null;
-  } catch {
-    return null;
-  }
-}
 
 /**
  * Checks whether the given config has enough fields to attempt a database connection.
